+++ FtcRobotController/src/main/java/org/firstinspires/ftc/robotcontroller/external/samples/ConceptScanServo.java
@@ -73,8 +73,8 @@ public class ConceptScanServo extends LinearOpMode {
         // Wait for the start button
         telemetry.addData(">", "Press Start to scan Servo." );
         telemetry.update();
-
         waitForStart();
+
 
         // Scan servo till stop pressed.
         while(opModeIsActive()){
