+++ TeamCode/build.gradle
@@ -20,5 +20,4 @@ repositories {
 
 dependencies {
     annotationProcessor files('lib/OpModeAnnotationProcessor.jar')
-    implementation 'org.openftc:easyopencv:1.4.4'
 }
