+++ TeamCode/build.release.gradle
@@ -5,6 +5,4 @@ dependencies {
     implementation 'org.firstinspires.ftc:FtcCommon:6.1.1'
     implementation (name: 'tfod-release', ext:'aar')
     implementation (name: 'tensorflow-lite-0.0.0-nightly', ext:'aar')
-
-    implementation 'com.acmerobotics.roadrunner:core:0.5.1'
 }
